third_party/WebKit/LayoutTests/fast/events/touch/gesture/pad-gesture-fling-expected.txt

Tests basic use of GestureFlingStart

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS cumulativeScrollX is >= 200
PASS cumulativeScrollY is >= 200
PASS successfullyParsed is true

TEST COMPLETE
